Revert "feat(vrl): add `parse_dnstap` function (#21985)"
This reverts commit 792aeeae5c216288bc11b78185ecfb7184ac3ece.
