+++ .github/actions/spelling/patterns.txt
@@ -226,7 +226,7 @@ user:P@ssw0rd
 /.+@base64/.+
 
 # Ignore base64 encoded values in VRL examples (requires padding to avoid false positives)
-"[A-Za-z0-9+\/]*=="
+"[A-Za-z0-9]*=="
 
 # ignore uuid_from_friendly_id argument
 uuid_from_friendly_id!\(".*"\)
