--- changelog.d/21985_parse_dnstap.feature.md
@@ -1,3 +1,0 @@
-Add VRL function `parse_dnstap` that can parse dnstap data and produce output in the same format as `dnstap` source.
-
-authors: esensar
