+++ lib/dnsmsg-parser/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT"
 
 [dependencies]
 data-encoding = "2.6"
-hickory-proto.workspace = true
+hickory-proto = { version = "0.24", features = ["dnssec"] }
 snafu.workspace = true
 
 [dev-dependencies]
