+++ src/conditions/vrl.rs
@@ -44,11 +44,7 @@ impl ConditionalConfig for VrlConfig {
 
         let functions = vrl::stdlib::all()
             .into_iter()
-            .chain(vector_lib::enrichment::vrl_functions());
-        #[cfg(feature = "sources-dnstap")]
-        let functions = functions.chain(dnstap_parser::vrl_functions());
-
-        let functions = functions
+            .chain(vector_lib::enrichment::vrl_functions())
             .chain(vector_vrl_functions::all())
             .collect::<Vec<_>>();
