+++ src/internal_events/dnstap.rs
@@ -24,3 +24,20 @@ impl<E: std::fmt::Display> InternalEvent for DnstapParseError<E> {
         .increment(1);
     }
 }
+
+#[derive(Debug)]
+pub(crate) struct DnstapParseWarning<E> {
+    pub error: E,
+}
+
+impl<E: std::fmt::Display> InternalEvent for DnstapParseWarning<E> {
+    fn emit(self) {
+        warn!(
+            message = "Recoverable error occurred while parsing dnstap data.",
+            error = %self.error,
+            stage = error_stage::PROCESSING,
+            error_type = error_type::PARSER_FAILED,
+            internal_log_rate_limit = true,
+        );
+    }
+}
